Add logging after reading the Data
