+++ ETL_Airflow/dags/metamorph_etl_dag.py
@@ -1,7 +1,7 @@
 from airflow.decorators import dag
 from datetime import datetime
-from tasks.ingestion_task import m_ingest_data_into_suppliers,m_ingest_data_into_products,m_ingest_data_into_customers,m_ingest_data_into_sales,m_load_supplier_performance
-
+from tasks.ingestion_task import m_ingest_data_into_suppliers,m_ingest_data_into_products,m_ingest_data_into_customers,m_ingest_data_into_sales
+from tasks.m_supplier_performance_task import m_load_supplier_performance
 
 @dag(
     dag_id="ingestion_data_pipeline",
@@ -14,7 +14,7 @@ def etl_process():
     product_task = m_ingest_data_into_products()
     customer_task = m_ingest_data_into_customers()
     sales_task = m_ingest_data_into_sales()
-    supplier_performance_task =m_load_supplier_performance()
+    supplier_performance_task = m_load_supplier_performance()
 
     # Set dependencies inside the DAG function
     [ supplier_task , product_task , customer_task , sales_task ]  >> supplier_performance_task
